added abstraced files. need to implement randomising solver

+++ Countdown/src/NumberNode.java
@@ -9,9 +9,9 @@
  */
 public class NumberNode extends SyntaxTreeNode
     {
-        public double Value;
+        public int Value;
 
-        public double Evaluate()
+        public int Evaluate()
         {
             return Value;
         }
